Move all script to etc/scripts
